refactor: reorganize page-specific components into page directories
Move page-specific components to colocate with their parent pages:

src/pages/
├── auth/
│   ├── LoginPage.tsx
│   ├── SignupPage.tsx
│   └── index.ts
├── business-plan/
│   ├── BusinessPlanViewer.tsx
│   ├── DiffView.tsx
│   ├── ExportDropdown.tsx
│   ├── SectionRegenerateButton.tsx
│   ├── VersionHistoryPanel.tsx
│   ├── VersionListItem.tsx
│   └── index.ts
├── calculator/
│   ├── FinancialCalculatorPage.tsx
│   ├── CTABanner.tsx
│   ├── PreviewFinancialForm.tsx
│   └── index.ts
├── profile/
│   ├── ProfilePage.tsx
│   ├── ProfileEditForm.tsx
│   ├── ChangePasswordForm.tsx
│   ├── DeleteAccountModal.tsx
│   └── index.ts
├── wizard/
│   ├── WizardStep.tsx
│   ├── FinancialSimulation.tsx
│   ├── PMFSurvey.tsx
│   ├── QuestionForm.tsx
│   └── index.ts
├── ProjectCreate.tsx
├── NotFoundPage.tsx
└── ServerErrorPage.tsx

src/components/ now contains only truly shared components:
├── auth/         - Shared auth components (AuthFormLayout, PasswordInput)
├── financial/    - Shared financial components (FinancialMetrics, Charts)
├── ui/           - UI primitives (Button, Input, etc.)
├── Modal/        - Modal components
├── Toast/        - Toast notifications
├── Layout.tsx
├── GlobalErrorBoundary.tsx
├── PageLoadingFallback.tsx
├── ComponentLoadingFallback.tsx
├── ProtectedRoute.tsx
└── SaveIndicator.tsx

Benefits:
- Page-specific code is colocated with pages
- Easier to understand component ownership
- Clearer separation of shared vs page-specific code
- Better code navigation

Build Status: ✅ Successful
