--- src/components/document/index.ts
@@ -1,14 +1,0 @@
-/**
- * 파일명: index.ts
- * 
- * 파일 용도:
- * document 컴포넌트 barrel export
- * 
- * 포함 컴포넌트:
- * - ExportDropdown: 내보내기 드롭다운
- * - SectionRegenerateButton: 섹션 재생성 버튼
- */
-
-export { ExportDropdown } from '../ExportDropdown';
-export { SectionRegenerateButton } from '../SectionRegenerateButton';
-
