--- src/components/profile/index.ts
@@ -1,16 +1,0 @@
-/**
- * 파일명: index.ts
- * 
- * 파일 용도:
- * profile 컴포넌트 barrel export
- * 
- * 포함 컴포넌트:
- * - ProfileEditForm: 프로필 수정 폼
- * - ChangePasswordForm: 비밀번호 변경 폼
- * - DeleteAccountModal: 회원 탈퇴 모달
- */
-
-export { ProfileEditForm } from '../ProfileEditForm';
-export { ChangePasswordForm } from '../ChangePasswordForm';
-export { DeleteAccountModal } from '../DeleteAccountModal';
-
