+++ src/pages/auth/LoginPage.tsx
@@ -16,9 +16,9 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import { LogIn, Mail, Lock, Eye, EyeOff, Sparkles } from 'lucide-react';
-import { useAuthStore } from '../stores/useAuthStore';
-import { useToast } from '../components/Toast';
-import { cn } from '../lib/utils';
+import { useAuthStore } from '../../stores/useAuthStore';
+import { useToast } from '../../components/Toast';
+import { cn } from '../../lib/utils';
 
 // 폼 스키마
 const loginSchema = z.object({
