+++ src/pages/auth/index.ts
@@ -1,0 +1,10 @@
+/**
+ * 파일명: index.ts
+ * 
+ * 파일 용도:
+ * auth 페이지 모듈 barrel export
+ */
+
+export { LoginPage } from './LoginPage';
+export { SignupPage } from './SignupPage';
+
