+++ src/pages/business-plan/index.ts
@@ -1,0 +1,15 @@
+/**
+ * 파일명: index.ts
+ * 
+ * 파일 용도:
+ * business-plan 페이지 모듈 barrel export
+ */
+
+export { BusinessPlanViewer } from './BusinessPlanViewer';
+export { VersionHistoryPanel } from './VersionHistoryPanel';
+export { VersionListItem } from './VersionListItem';
+export type { DocumentVersion, VersionStatus } from './VersionListItem';
+export { DiffView } from './DiffView';
+export { SectionRegenerateButton } from './SectionRegenerateButton';
+export { ExportDropdown } from './ExportDropdown';
+
