+++ src/pages/calculator/CTABanner.tsx
@@ -13,8 +13,8 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Sparkles, ArrowRight, Lock, Zap, Shield, Save } from 'lucide-react';
-import { Button } from './ui';
-import { cn } from '../lib/utils';
+import { Button } from '../../components/ui';
+import { cn } from '../../lib/utils';
 
 interface CTABannerProps {
   /** 배너 변형 */
