+++ src/pages/profile/DeleteAccountModal.tsx
@@ -13,9 +13,9 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { AlertTriangle, Lock, Eye, EyeOff, Loader2, X } from 'lucide-react';
-import { useAuthStore } from '../stores/useAuthStore';
-import { useToast } from './Toast';
-import { cn } from '../lib/utils';
+import { useAuthStore } from '../../stores/useAuthStore';
+import { useToast } from '../../components/Toast';
+import { cn } from '../../lib/utils';
 
 interface DeleteAccountModalProps {
   /** 모달 열림 상태 */
