+++ src/pages/profile/index.ts
@@ -1,0 +1,12 @@
+/**
+ * 파일명: index.ts
+ * 
+ * 파일 용도:
+ * profile 페이지 모듈 barrel export
+ */
+
+export { ProfilePage } from './ProfilePage';
+export { ProfileEditForm } from './ProfileEditForm';
+export { ChangePasswordForm } from './ChangePasswordForm';
+export { DeleteAccountModal } from './DeleteAccountModal';
+
