+++ src/pages/wizard/QuestionForm.tsx
@@ -33,7 +33,7 @@ import React from 'react';
 import { useWizardStore } from '../../stores/useWizardStore';
 import { useAutoSave } from '../../hooks/useAutoSave';
 import { Question } from '../../types';
-import { Input, Textarea } from '../ui';
+import { Input, Textarea } from '../../components/ui';
 
 interface QuestionFormProps {
   /** 렌더링할 질문 목록 */
