+++ src/pages/wizard/index.ts
@@ -1,0 +1,12 @@
+/**
+ * 파일명: index.ts
+ * 
+ * 파일 용도:
+ * wizard 페이지 모듈 barrel export
+ */
+
+export { WizardStep } from './WizardStep';
+export { FinancialSimulation } from './FinancialSimulation';
+export { PMFSurvey } from './PMFSurvey';
+export { QuestionForm } from './QuestionForm';
+
